% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ape}
\alias{ape}
\title{Evaluation of absolute prediction errors.}
\usage{
ape(x, surv, time, censored = "conditional")
}
\arguments{
\item{x}{A vector with the predictions in [0,1]}

\item{surv}{The Surv() object of the observed survival.}

\item{time}{The time at which the predictions are to be evaluated.}

\item{censored}{A character vector indicating the strategy to handle cases censored before time.}
}
\value{
A vector of length
}
\description{
Given predictions p in [0,1], the function evaluates the mean absolute prediction error sum(|p-status|)/n, the mean squared error (Brier score)
sum((p-status)^2)/n, the log2 entropy sum(p * log2(status/p))/n, and a Bayes misclassification rate sum((p>.5)*(1-status))/n. If the
status is extrapolated with probability in [0,1], weighted averages are reported.
}
\author{
mg14
}

