% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ErrorControlCPSS}
\alias{ErrorControlCPSS}
\title{Type-1 error control for CoxCPSS}
\usage{
ErrorControlCPSS(coxCPSS, control = coxCPSS$control, level = coxCPSS$level,
  pi.thr = coxCPSS$pi.thr, which.error = coxCPSS$which.error)
}
\arguments{
\item{coxCPSS}{A CoxCPSS object}

\item{control}{Which type of control}

\item{level}{The level for P-value adjustments}

\item{pi.thr}{The threshold for stable variables}

\item{which.error}{Which covariates are used for estimating the parameter of the null-distribution}
}
\value{
CoxCPSS
}
\description{
This function implements P-value based error control for the CPSS (Shah and Samworth, 2013). This transforms the selection probability into P-values
for L-concave distributions. Common multiple testing adjustments can then be performed on those P-values
}
\author{
mg14
}

