% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MakeInteger}
\alias{MakeInteger}
\title{Convert factor to integer.}
\usage{
MakeInteger(F)
}
\arguments{
\item{F}{A factor}
}
\value{
A data.frame() with columns corresponding to levels() in the factor.
}
\description{
Convert factor to integer.
}
\author{
mg14
}

