\name{PartialC}
\alias{PartialC}
\title{Compute concordance for risk components}
\usage{
  PartialC(fit, newX = NULL, newSurv = NULL)
}
\arguments{
  \item{fit}{}

  \item{newX}{}

  \item{newSurv}{}
}
\value{
  A vector with the concordance of each component
}
\description{
  Compute concordance for risk components
}
\author{
  mg14
}

