\name{PartialRiskVar}
\alias{PartialRiskVar}
\title{Variance (confidence intervals) of partial risk components}
\usage{
  PartialRiskVar(fit, newX = NULL)
}
\arguments{
  \item{fit}{}

  \item{newX}{}
}
\value{
  A matrix with the confidence interval (prediction
  variance) for each risk component
}
\description{
  Variance (confidence intervals) of partial risk
  components
}
\author{
  mg14
}

