% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.CoxCPSS}
\alias{plot.CoxCPSS}
\title{Plot a CoxCPSS model}
\usage{
\method{plot}{CoxCPSS}(x, xlab = "1/lambda", ylab = "Selection probability",
  lty = rep(1, ncol(x$X)), log = "", col = c("black", "grey"),
  xlim = range(1/x$lambda[x$Lambda]), ylim = c(0, 1), ...)
}
\arguments{
\item{x}{a CoxCPSS model}

\item{xlab}{The x-label}

\item{ylab}{The y-label}

\item{lty}{Line type}

\item{log}{Whether to use a log scale}

\item{col}{The colors}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{...}{Additional options passed to plot()}
}
\description{
Plot a CoxCPSS model
}
\author{
mg14
}

