% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.CoxRFX}
\alias{plot.CoxRFX}
\title{Plot a CoxRFX model}
\usage{
\method{plot}{CoxRFX}(x, ...)
}
\arguments{
\item{x}{The CoxRFX object}

\item{...}{Additional parameters passed to plot(). This can include the non-standard argument order (=1:nlevels(x$groups)), to reorder the groups.}
}
\description{
This plots a CoxRFX model
}
\author{
mg14
}

