\name{predict.CoxCPSS}
\alias{predict.CoxCPSS}
\title{Predict method for a CoxCPSS object}
\usage{
  predict.CoxCPSS(x, ...)
}
\arguments{
  \item{x}{A CoxCPSS fit}

  \item{...}{Parameters passed on to predict.coxph}
}
\value{
  Depending on the arguments ... either a vector of the log
  hazard ratio, or something else.
}
\description{
  Predict method for a CoxCPSS object
}
\author{
  mg14
}

