% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.CoxRFX}
\alias{print.CoxRFX}
\title{Print method for CoxRFX}
\usage{
\method{print}{CoxRFX}(x, ...)
}
\arguments{
\item{x}{CoxRFX}

\item{...}{Currently unused}
}
\description{
This function implicitly calls summary.CoxRFX().
}
\author{
mg14
}

