% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SimDataNonp}
\alias{SimDataNonp}
\title{Non-parametric data extrapolations based on multiple imputation
A fraction percentMissing is set to NA and multiple imputation from the mice package is used to fill these gaps.}
\usage{
SimDataNonp(X, nData, percentMissing = 0.33, ...)
}
\arguments{
\item{X}{The original data. A matrix or data.frame.}

\item{nData}{The number of new data points (rows).}

\item{percentMissing}{Fraction of missing values. Default .33.}

\item{...}{Additional arguments passed to mice.}
}
\value{
A data.frame with extrapolated data
}
\description{
Non-parametric data extrapolations based on multiple imputation
A fraction percentMissing is set to NA and multiple imputation from the mice package is used to fill these gaps.
}
\author{
mg14
}

