\name{SimSurvNonp}
\alias{SimSurvNonp}
\title{Non-parametric survival simulations}
\usage{
  SimSurvNonp(risk, H0, surv, ...)
}
\arguments{
  \item{risk}{A vector of the relative risk/}

  \item{H0}{The baseline hazed function, typically derived
  from basehaz().}

  \item{surv}{A survival object}

  \item{...}{}
}
\value{
  A Surv()ival object
}
\description{
  Non-parametric survival simulations
}
\author{
  mg14
}

