% Generated by roxygen2 (4.0.1): do not edit by hand
\name{StandardizeMagnitude}
\alias{StandardizeMagnitude}
\title{Standardize the magnitude of covariates}
\usage{
StandardizeMagnitude(X)
}
\arguments{
\item{X}{}
}
\value{
data.frame of dim(X)
}
\description{
Standardize the magnitude of covariates
}
\author{
mg14
}

