% Generated by roxygen2 (4.0.1): do not edit by hand
\name{survStatus}
\alias{survStatus}
\title{Extract and predict survival status}
\usage{
survStatus(surv, time, censored = c("conditional", "na"))
}
\arguments{
\item{surv}{As Surv() object. Currently only right-censoring supported.}

\item{time}{The time at which to extract the survival status.}

\item{censored}{A character vector, either "conditional" or "na".}
}
\value{
A vector with the survival status.
}
\description{
This function extracts the survival status at a given time t. For censored patients at time t_cens < t, the
predicted status is S[t]/S[t_cens] if censored="conditional" and NA otherwise.
}
\author{
mg14
}

