% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TestInteractionsCPH}
\alias{TestInteractionsCPH}
\title{Test for interactions
A coxph() model is used to systematically test for interactions. Returned are both a Wald test and a likelihood ratio test,
as well as the estimated coefficient and an indicator if the fitting caused a warning().}
\usage{
TestInteractionsCPH(data, survival, pairs, whichMain = colnames(data),
  mc.cores = 1, minObs = 5)
}
\arguments{
\item{data}{A data.frame() with covariates}

\item{survival}{A Surv()ival object}

\item{pairs}{A list() with interactions to be considered}

\item{whichMain}{A vector() of main effects to include. The main effects of the pair of interests are always included/}

\item{minObs}{The minimal number of observations for main effects other than the interaction.}

\item{mc.cores}{The number of cores to use.}
}
\value{
A data.frame() with columns pWald, pLR, coef and warn.
}
\description{
Test for interactions
A coxph() model is used to systematically test for interactions. Returned are both a Wald test and a likelihood ratio test,
as well as the estimated coefficient and an indicator if the fitting caused a warning().
}
\author{
mg14
}

