\name{VarianceComponents}
\alias{VarianceComponents}
\title{Variance components}
\usage{
  VarianceComponents(fit, type = c("diag", "colSums"))
}
\arguments{
  \item{fit}{}

  \item{type}{}
}
\value{
  A vector containing the variance components
}
\description{
  Variance components
}
\author{
  mg14
}

