% Generated by roxygen2 (4.0.1): do not edit by hand
\name{VarianceComponentsCI}
\alias{VarianceComponentsCI}
\title{Confidence intervals of variance components}
\usage{
VarianceComponentsCI(fit, newZ = fit$Z, groups = fit$groups, q = c(0.025,
  0.5, 0.975), n = 200, mc.cores = 1)
}
\arguments{
\item{fit}{The CoxRFX model}

\item{newZ}{The Z (data matrix). Default = fit$Z}

\item{groups}{The groups. Default fit$groups}

\item{q}{The quantiles to be evaluated. Default q = c(0.275, 0.5, 0.975), corresponding to the median and symmetric 95% confidence intervals.}

\item{n}{The number of samples. Default = 200.}

\item{mc.cores}{The number of mc.cores to use. Default=1.}
}
\value{
A numeric matrix with summary statistics.
}
\description{
This function numerically computes confidence intervals for the variance components, based sampling on the variances of individual effects.
}
\author{
mg14
}

