% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CoxCPSS}
\alias{CoxCPSS}
\title{Complementary pairs stability selection for the Cox proportional hazards model}
\usage{
CoxCPSS(X, surv, bootstrap.samples = 50, nlambda = 250, alpha.weak = 0.5,
  penalty.factor = rep(1, ncol(X)), mc.cores = 1, pi.thr = 0.8,
  control = c("BH", "theta", "FDR", p.adjust.methods), level = 0.1,
  simultaneous = TRUE, which.error = 1:ncol(X), keep.trying = TRUE,
  seed = 42, coxph = TRUE)
}
\arguments{
\item{X}{matrix or data.frame of the covariats}

\item{surv}{a Surv() object}

\item{bootstrap.samples}{Number of bootstrap samples (default = 50). Note: These give 100 samples for CPSS.}

\item{nlambda}{Approximate number of datapoints for the LASSO penalty. Default 250.}

\item{alpha.weak}{Weakness parameter for the stability selection. Default 0.5.}

\item{penalty.factor}{Penalty factor. Allows for reducing the penalty on some covariates. Default rep(1,ncol(X)).}

\item{mc.cores}{Number of cores if parallel computations are desired}

\item{pi.thr}{The stability threshold. Default = 0.8.}

\item{control}{What type of error control is desired. Use "BH" for Benjamini-Hochberg correction.}

\item{level}{The level for P-value adjustments.}

\item{simultaneous}{TRUE for complementary pairs stability selection. Needed for type-1 error control}

\item{which.error}{Indeces of covariates used for estimating the selection probability under the null. Default = 1:ncol(X).}

\item{keep.trying}{Sometimes the glmnet algorithm fails to compute the entire LASSO trace. If keep.trying = TRUE (default) the procedure will continue until 'bootstrap.samples' complete solutions were completed.}

\item{seed}{The seed of the random number generator. Set to 42 by default to assure reproducibility. Set to NULL if you don't want it to be set.}

\item{coxph}{Wether to fit a coxph model using the selected variables.}
}
\value{
CoxCPSS
}
\description{
Complementary pairs stability selection for the Cox proportional hazards model
}
\author{
mg14
}
\references{
R. D. Shah and R. J. Samworth (2013). Variable selection with error control: another look at stability selection. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 75:55--80. http://dx.doi.org/10.1111/j.1467-9868.2011.01034.x

N. Meinshausen and P. Bühlmann (2010). Stability selection. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 72:417--473. http://dx.doi.org/10.1111/j.1467-9868.2010.00740.x
}

