\name{CoxCPSSInteractions}
\alias{CoxCPSSInteractions}
\title{CoxCPSS with interaction terms}
\usage{
  CoxCPSSInteractions(X, surv, scope = 1:ncol(X), ...)
}
\arguments{
  \item{X}{}

  \item{surv}{}

  \item{scope}{}

  \item{...}{}
}
\value{
  CoxCPSS
}
\description{
  CoxCPSS with interaction terms
}
\author{
  mg14
}

