% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PartialC}
\alias{PartialC}
\title{Compute concordance for risk components}
\usage{
PartialC(fit, newX = fit$X, newSurv = fit$surv, groups = fit$groups)
}
\arguments{
\item{fit}{The CoxRFX fit}

\item{newX}{New data, defaults to fit$X}

\item{groups}{The groups, defaults to fit$groups}

\item{newSurv}{The survival object, defaults to fit$surv}
}
\value{
A vector with the concordance of each component
}
\description{
Compute concordance for risk components
}
\author{
mg14
}

