% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.CoxRFX}
\alias{plot.CoxRFX}
\title{Plot a CoxRFX model}
\usage{
\method{plot}{CoxRFX}(fit, col = c(brewer.pal(9, "Set1"), brewer.pal(8,
  "Dark2")), order = 1:nlevels(fit$groups), xlim = range(coef(fit)),
  xlab = "Coefficient", ...)
}
\arguments{
\item{fit}{}

\item{col}{}

\item{order}{}

\item{xlim}{}

\item{xlab}{}

\item{...}{}
}
\description{
Plot a CoxRFX model
}
\author{
mg14
}

