% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PlotVarianceComponents}
\alias{PlotVarianceComponents}
\title{Plot variance components}
\usage{
PlotVarianceComponents(fit, col = 1:nlevels(fit$groups),
  groups = fit$groups, type = "rowSums")
}
\arguments{
\item{fit}{The CoxRFX fit}

\item{col}{The colors for each component}

\item{groups}{the groups to be used, if different from the fitted ones.}

\item{type}{The type of variance compnents: Either 'rowSums' (default) or 'diag'. Rowsums sum up to the actual variance of the linear predictor, but can be
negative. Plotting just the 'diag'onal elements guarantees positive components}
}
\description{
Plot variance components
}
\author{
mg14
}

