% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.CoxCPSS}
\alias{predict.CoxCPSS}
\title{Predict method for a CoxCPSS object}
\usage{
\method{predict}{CoxCPSS}(object, ...)
}
\arguments{
\item{object}{A CoxCPSS fit}

\item{...}{Parameters passed on to predict.coxph}
}
\value{
Depending on the arguments ... either a vector of the log hazard ratio, or something else.
}
\description{
Predict method for a CoxCPSS object
}
\author{
mg14
}

